Add Event to Notify ASL Script Change
This will allow other components that use the ASLScript introspection (like ASLVarViewer) to more easily detect when a Script changes.  This allows for a more defined life-cycle rather than trying to connect at certain events like start or end of run.

This would enable zoton2's scenario to better enable using ASLVarViewer in the future for 1.7 as a debugger to display script values in LiveSplit vs. using dbgview print statements by allowing it to update its reference as soon as a script change is made and detected by LiveSplit.